feat(ui): move theme control to Settings, remove from nav
