+++ src/routes/(app)/settings/+page.svelte
@@ -16,6 +16,7 @@
 	import { Pencil, Plus, RotateCcw } from '@lucide/svelte';
 	import { getContext } from 'svelte';
 	import { t, getLocale, SUPPORTED_LOCALES, LOCALE_LABELS } from '$lib/i18n';
+	import { applyTheme, saveTheme, THEMES, THEME_ICONS, type Theme } from '$lib/theme';
 
 	let { data, form }: { data: PageData; form: ActionData } = $props();
 
@@ -24,6 +25,15 @@
 
 	let showPasswordFields = $state(false);
 	let localeForm: HTMLFormElement;
+
+	let themeOverride = $state<Theme | null>(null);
+	let currentTheme = $derived<Theme>(themeOverride ?? (data.user?.theme as Theme) ?? 'system');
+
+	async function setTheme(theme: Theme) {
+		themeOverride = theme;
+		applyTheme(theme);
+		await saveTheme(theme);
+	}
 
 	const browserOrigin = $derived(typeof window !== 'undefined' ? window.location.origin : '');
 
@@ -225,6 +235,32 @@
 					</Select>
 				</div>
 			</form>
+		</CardContent>
+	</Card>
+
+	<Card>
+		<CardHeader>
+			<CardTitle>{t(locale, 'page.settings.appearanceCard')}</CardTitle>
+		</CardHeader>
+		<CardContent>
+			<div class="flex rounded-md border border-border p-0.5 gap-0.5 bg-muted">
+				{#each THEMES as theme (theme)}
+					{@const Icon = THEME_ICONS[theme]}
+					<button
+						type="button"
+						onclick={() => setTheme(theme)}
+						aria-label="{theme.charAt(0).toUpperCase() + theme.slice(1)} mode"
+						aria-pressed={currentTheme === theme}
+						class="flex-1 flex items-center justify-center gap-1.5 rounded px-3 py-2 text-sm transition-all {currentTheme ===
+						theme
+							? 'bg-background text-foreground shadow-sm'
+							: 'text-muted-foreground hover:text-foreground'}"
+					>
+						<Icon class="h-4 w-4 shrink-0" />
+						<span>{theme.charAt(0).toUpperCase() + theme.slice(1)}</span>
+					</button>
+				{/each}
+			</div>
 		</CardContent>
 	</Card>
